+++ remote/src/main/java/kr/ninety/remote/di/DataSourceModule.kt
@@ -4,15 +4,36 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import kr.ninety.data.remote.RemoteDataSource
-import kr.ninety.remote.RemoteDataSourceImpl
+import kr.ninety.data.remote.RemoteAlbumDataSource
+import kr.ninety.data.remote.RemoteFilmDataSource
+import kr.ninety.data.remote.RemotePhotoDataSource
+import kr.ninety.data.remote.RemoteUserDataSource
+import kr.ninety.remote.RemoteAlbumDataSourceImpl
+import kr.ninety.remote.RemoteFilmDataSourceImpl
+import kr.ninety.remote.RemotePhotoDataSourceImpl
+import kr.ninety.remote.RemoteUserDataSourceImpl
 
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class DataSourceModule {
 
     @Binds
-    abstract fun provideRemoteDataSource(
-        remoteDataSourceImpl: RemoteDataSourceImpl
-    ): RemoteDataSource
+    abstract fun provideRemoteUserDataSource(
+        remoteUserDataSourceImpl: RemoteUserDataSourceImpl
+    ): RemoteUserDataSource
+
+    @Binds
+    abstract fun provideRemoteFilmDataSource(
+        remoteFilmDataSourceImpl: RemoteFilmDataSourceImpl
+    ): RemoteFilmDataSource
+
+    @Binds
+    abstract fun provideRemotePhotoDataSource(
+        remotePhotoDataSourceImpl: RemotePhotoDataSourceImpl
+    ): RemotePhotoDataSource
+
+    @Binds
+    abstract fun provideRemoteAlbumDataSource(
+        remoteAlbumDataSourceImpl: RemoteAlbumDataSourceImpl
+    ): RemoteAlbumDataSource
 }
